Verify the engine headlessly, and rebalance the cat's day
Added Tools/linux-verify: hand-written stand-ins for CoreGraphics, QuartzCore,
UIKit, SceneKit, AVFoundation, UserNotifications and SwiftUI, so the whole
project compiles and runs on any Swift toolchain. One script builds the shims,
stages the sources, and runs ~13M assertions across maths, the solar clock,
mesh generation, the rig, the animator, the brain, interaction, consumables,
offline catch-up, the save file, room layout, textures and lighting. A second
mode simulates whole days and reports how the cat spends them.

Compile errors this found and fixed:
- SCNAction.rotateBy takes CGFloat, not the Float that deg() returns
- CACurrentMediaTime needs an explicit QuartzCore import
- a Float/CGFloat mix in the coat texture's belly gradient

Behavioural bugs this found and fixed:
- Scoring noise was re-rolled on every comparison, including for the current
  activity, so the cat dithered. Jitter is now held for 45s at a time.
- Re-evaluation ran every 2.5s regardless of activity, waking sleeping cats
  constantly. The interval now scales with the activity's length, and urgent
  needs can still cut it short.
- Waking from a nap always triggered a stretch, which fed straight back into
  sleeping. Stretch now needs a real nap behind it and carries a cooldown; it
  was taking up to 21% of the day and is now a transition again.
- Rest was restored roughly 70x faster than it drained, so sleep was nearly
  free. Sleeping is now hours of work, sleep pressure follows the rest deficit
  rather than a constant floor, and rest drain follows sleepiness as well as
  energy so a Chill cat sleeps more than a Hunter, not less.
- greetPlayer, sitAndStare, followPlayer and chirpAtBirds could never win a
  scoring round, so the cat never sought the player out at all.
- The feeder and fountain emptied in about a day; they now last most of a week.
- The cat vocalised over 1400 times a day. Now 120-200, about one every ten
  minutes.

The result: 42-53% of the day asleep depending on temperament, 25 distinct
activities in play, and a different-looking day for each archetype.

+++ MeowRoom/Scene/GameSceneController.swift
@@ -1,5 +1,6 @@
 import Foundation
 import SceneKit
+import QuartzCore
 import UIKit
 
 /// Owns the SceneKit scene and drives it from the simulation every frame.
@@ -351,7 +352,7 @@ final class GameSceneController: NSObject, SCNSceneRendererDelegate {
         teacupFalling = true
         let fall = SCNAction.group([
             SCNAction.move(by: SCNVector3(x: 0.18, y: -RoomLayout.tableTop, z: 0.10), duration: 0.55),
-            SCNAction.rotateBy(x: deg(120), y: deg(40), z: deg(90), duration: 0.55)
+            SCNAction.rotateBy(x: CGFloat(deg(120)), y: CGFloat(deg(40)), z: CGFloat(deg(90)), duration: 0.55)
         ])
         fall.timingMode = .easeIn
         let restore = SCNAction.sequence([
